requirements.txt
pymysql


django
djangorestframework
djangorestframework-simplejwt